MAINT: refactor Bessel to basis module
